se saca el telegram de Program
