+++ src/Library/Library.csproj
@@ -4,4 +4,9 @@
     <TargetFramework>netcoreapp3.1</TargetFramework>
   </PropertyGroup>
 
+  <ItemGroup>
+    <PackageReference Include="Telegram.Bot" Version="15.5.1" />
+    <PackageReference Include="Telegram.Bot.Extensions.Polling" Version="0.1.3" />
+  </ItemGroup>
+
 </Project>
