--- src/Library/Plataform/TelegramPlataform.cs
@@ -1,1 +1,0 @@
-
