+++ src/Test/Library.Test/TestConfiguration.cs
@@ -1,0 +1,28 @@
+using NUnit.Framework;
+using Library;
+
+namespace Library.Test
+{
+    public class TestConfig
+    {
+        
+        
+        [Test]
+        public void NumberOfConfigurationListElementMustBeThree()
+        {
+            Configuration configurationTest = new Configuration("Configuration.csv");
+            int excpected = 3;
+            
+            List[] list = configurationTest.Paraments();
+            int result = 0;
+            foreach(var elem in list)
+            {
+                result += 1;
+            }
+            Assert.AreEqual(result, excpected);
+
+        }
+
+        
+    }
+}
